requirements.txt
Flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.1.0
requests==2.32.3
